added FindInactiveGameObject methods to extensions class
